feat: add auto login for bad user sessions
