+++ cli/packages/cmd/root.go
@@ -77,6 +77,7 @@ func init() {
 
 	isTelemetryOn, _ := rootCmd.PersistentFlags().GetBool("telemetry")
 	Telemetry = telemetry.NewTelemetry(isTelemetryOn)
+	util.Telemetry = Telemetry
 }
 
 func initLog() {
